Separate out docker interface and add file to codecov ignore
